Split the Block device in multiple files.
- mod.rs contains constant defs and errors
- device.rs will be the Block device struct, helper
functions and VirtioDevice trait implementation
- event_handler.rs implements the EventHandler trait
- request.rs holds Virtio Block Request implementation
