+++ src/devices/src/virtio/block/request.rs
@@ -1,0 +1,225 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+//
+// Portions Copyright 2017 The Chromium OS Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the THIRD-PARTY file.
+use std::convert::From;
+use std::io::{self, Read, Seek, SeekFrom, Write};
+use std::result;
+
+use logger::{Metric, METRICS};
+use memory_model::{DataInit, GuestAddress, GuestMemory, GuestMemoryError};
+use virtio_gen::virtio_blk::*;
+
+use super::super::DescriptorChain;
+use super::{Error, SECTOR_SHIFT, SECTOR_SIZE};
+
+#[derive(Debug)]
+pub enum ExecuteError {
+    BadRequest(Error),
+    Flush(io::Error),
+    Read(GuestMemoryError),
+    Seek(io::Error),
+    Write(GuestMemoryError),
+    Unsupported(u32),
+}
+
+impl ExecuteError {
+    pub fn status(&self) -> u32 {
+        match *self {
+            ExecuteError::BadRequest(_) => VIRTIO_BLK_S_IOERR,
+            ExecuteError::Flush(_) => VIRTIO_BLK_S_IOERR,
+            ExecuteError::Read(_) => VIRTIO_BLK_S_IOERR,
+            ExecuteError::Seek(_) => VIRTIO_BLK_S_IOERR,
+            ExecuteError::Write(_) => VIRTIO_BLK_S_IOERR,
+            ExecuteError::Unsupported(_) => VIRTIO_BLK_S_UNSUPP,
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub enum RequestType {
+    In,
+    Out,
+    Flush,
+    GetDeviceID,
+    Unsupported(u32),
+}
+
+impl From<u32> for RequestType {
+    fn from(value: u32) -> Self {
+        match value {
+            VIRTIO_BLK_T_IN => RequestType::In,
+            VIRTIO_BLK_T_OUT => RequestType::Out,
+            VIRTIO_BLK_T_FLUSH => RequestType::Flush,
+            VIRTIO_BLK_T_GET_ID => RequestType::GetDeviceID,
+            t => RequestType::Unsupported(t),
+        }
+    }
+}
+
+pub struct Request {
+    pub request_type: RequestType,
+    pub data_len: u32,
+    pub status_addr: GuestAddress,
+    sector: u64,
+    data_addr: GuestAddress,
+}
+
+/// The request header represents the mandatory fields of each block device request.
+///
+/// A request header contains the following fields:
+///   * request_type: an u32 value mapping to a read, write or flush operation.
+///   * reserved: 32 bits are reserved for future extensions of the Virtio Spec.
+///   * sector: an u64 value representing the offset where a read/write is to occur.
+///
+/// The header simplifies reading the request from memory as all request follow
+/// the same memory layout.
+#[derive(Copy, Clone)]
+#[repr(C)]
+pub struct RequestHeader {
+    request_type: u32,
+    _reserved: u32,
+    sector: u64,
+}
+
+// Safe because RequestHeader only contains plain data.
+unsafe impl DataInit for RequestHeader {}
+
+impl RequestHeader {
+    /// Reads the request header from GuestMemory starting at `addr`.
+    ///
+    /// Virtio 1.0 specifies that the data is transmitted by the driver in little-endian
+    /// format. Firecracker currently runs only on little endian platforms so we don't
+    /// need to do an explicit little endian read as all reads are little endian by default.
+    /// When running on a big endian platform, this code should not compile, and support
+    /// for explicit little endian reads is required.
+    #[cfg(target_endian = "little")]
+    fn read_from(memory: &GuestMemory, addr: GuestAddress) -> result::Result<Self, Error> {
+        let request_header: RequestHeader = memory
+            .read_obj_from_addr(addr)
+            .map_err(Error::GuestMemory)?;
+        Ok(request_header)
+    }
+}
+
+impl Request {
+    pub fn parse(
+        avail_desc: &DescriptorChain,
+        mem: &GuestMemory,
+    ) -> result::Result<Request, Error> {
+        // The head contains the request type which MUST be readable.
+        if avail_desc.is_write_only() {
+            return Err(Error::UnexpectedWriteOnlyDescriptor);
+        }
+
+        let request_header = RequestHeader::read_from(mem, avail_desc.addr)?;
+        let mut req = Request {
+            request_type: RequestType::from(request_header.request_type),
+            sector: request_header.sector,
+            data_addr: GuestAddress(0),
+            data_len: 0,
+            status_addr: GuestAddress(0),
+        };
+
+        let data_desc;
+        let status_desc;
+        let desc = avail_desc
+            .next_descriptor()
+            .ok_or(Error::DescriptorChainTooShort)?;
+
+        if !desc.has_next() {
+            status_desc = desc;
+            // Only flush requests are allowed to skip the data descriptor.
+            if req.request_type != RequestType::Flush {
+                return Err(Error::DescriptorChainTooShort);
+            }
+        } else {
+            data_desc = desc;
+            status_desc = data_desc
+                .next_descriptor()
+                .ok_or(Error::DescriptorChainTooShort)?;
+
+            if data_desc.is_write_only() && req.request_type == RequestType::Out {
+                return Err(Error::UnexpectedWriteOnlyDescriptor);
+            }
+            if !data_desc.is_write_only() && req.request_type == RequestType::In {
+                return Err(Error::UnexpectedReadOnlyDescriptor);
+            }
+            if !data_desc.is_write_only() && req.request_type == RequestType::GetDeviceID {
+                return Err(Error::UnexpectedReadOnlyDescriptor);
+            }
+
+            req.data_addr = data_desc.addr;
+            req.data_len = data_desc.len;
+        }
+
+        // The status MUST always be writable.
+        if !status_desc.is_write_only() {
+            return Err(Error::UnexpectedReadOnlyDescriptor);
+        }
+
+        if status_desc.len < 1 {
+            return Err(Error::DescriptorLengthTooSmall);
+        }
+
+        req.status_addr = status_desc.addr;
+
+        Ok(req)
+    }
+
+    pub fn execute<T: Seek + Read + Write>(
+        &self,
+        disk: &mut T,
+        disk_nsectors: u64,
+        mem: &GuestMemory,
+        disk_id: &[u8],
+    ) -> result::Result<u32, ExecuteError> {
+        let mut top: u64 = u64::from(self.data_len) / SECTOR_SIZE;
+        if u64::from(self.data_len) % SECTOR_SIZE != 0 {
+            top += 1;
+        }
+        top = top
+            .checked_add(self.sector)
+            .ok_or(ExecuteError::BadRequest(Error::InvalidOffset))?;
+        if top > disk_nsectors {
+            return Err(ExecuteError::BadRequest(Error::InvalidOffset));
+        }
+
+        disk.seek(SeekFrom::Start(self.sector << SECTOR_SHIFT))
+            .map_err(ExecuteError::Seek)?;
+
+        match self.request_type {
+            RequestType::In => {
+                mem.read_to_memory(self.data_addr, disk, self.data_len as usize)
+                    .map_err(ExecuteError::Read)?;
+                METRICS.block.read_bytes.add(self.data_len as usize);
+                METRICS.block.read_count.inc();
+                return Ok(self.data_len);
+            }
+            RequestType::Out => {
+                mem.write_from_memory(self.data_addr, disk, self.data_len as usize)
+                    .map_err(ExecuteError::Write)?;
+                METRICS.block.write_bytes.add(self.data_len as usize);
+                METRICS.block.write_count.inc();
+            }
+            RequestType::Flush => match disk.flush() {
+                Ok(_) => {
+                    METRICS.block.flush_count.inc();
+                    return Ok(0);
+                }
+                Err(e) => return Err(ExecuteError::Flush(e)),
+            },
+            RequestType::GetDeviceID => {
+                if (self.data_len as usize) < disk_id.len() {
+                    return Err(ExecuteError::BadRequest(Error::InvalidOffset));
+                }
+                mem.write_slice_at_addr(disk_id, self.data_addr)
+                    .map_err(ExecuteError::Write)?;
+            }
+            RequestType::Unsupported(t) => return Err(ExecuteError::Unsupported(t)),
+        };
+        Ok(0)
+    }
+}
